requirements.txt
streamlit==1.51.0
crewai==0.30.0
chromadb==0.4.24
langchain==0.1.20
langchain-community==0.0.38
sentence-transformers==2.2.2
pandas==2.3.3
pdfplumber==0.11.7
python-dotenv==1.2.1
groq==0.36.0
pypdf==4.3.1
numpy==1.26.4
requests==2.32.5
